Add initial preparation for WASM support
It would be really cool to have a landing page where you can try out
RCL in your browser. And with WASM, this should be possible nowadays.

MDN has a page that mentions wasm-pack, but I would like to avoid the
Javascript/NPM ecosystem like the plague, and although wasm-pack looks
like it is safe, so far it also looks like we don't *need* it, so I
prefer to go for the lower-level approach.

+++ src/source.rs
@@ -68,7 +68,7 @@ impl Span {
         // an input larger than 262 TiB before it happens, so we can get away
         // with a panic.
         assert!(
-            end <= 0xffff_ffff_ffff,
+            end as u128 <= 0xffff_ffff_ffff,
             // coverage:off -- Error is not tested, I don't have such large disk.
             "Document should not be larger than 262 TiB.",
             // coverage:on
